fix(armv8/vgic): return the full 64-bit route on irouter reads
A doubleword read of IROUTER returned only the low word, truncating
Aff3.

+++ src/arch/armv8/vgicv3.c
@@ -155,7 +155,7 @@ static void vgicd_emul_router_access(struct emul_access* acc,
         } else if (word_access) {
             vcpu_writereg(cpu()->vcpu, acc->reg, (uint32_t)route);
         } else {
-            vcpu_writereg(cpu()->vcpu, acc->reg, (uint32_t)route);
+            vcpu_writereg(cpu()->vcpu, acc->reg, (unsigned long)route);
         }
     } else {
         uint64_t reg_value = vcpu_readreg(cpu()->vcpu, acc->reg);
